Add convenience function to create a reusable API client
CLI/SDK are on an old PineJS version which is incompatible with how we make requests — specifically, handling top-level options like `$filter` and `$expand`.

Given we already include PineJS as a dependency for its types, add a helper to create a compatible client that can be reused across requests.

Change-Type: minor